refactor badges
