+++ src/components/badges/BadgeBlock.tsx
@@ -1,9 +1,9 @@
 import { AccentText } from 'components/Text'
 import { useSnapshot } from 'valtio'
 import { useState } from 'react'
-import BadgeCard from 'components/BadgeCard'
+import BadgeCard from 'components/badges/BadgeCard'
 import BadgeIcon from 'icons/BadgeIcon'
-import BadgeWrapper from 'components/BadgeWrapper'
+import BadgeWrapper from 'components/badges/BadgeWrapper'
 import Button from 'components/Button'
 import Complete from 'icons/Complete'
 import ContractName from 'components/ContractName'
