+++ src/components/badges/BadgesOwnedForContractLoading.tsx
@@ -1,0 +1,28 @@
+import BadgeCard from 'components/badges/BadgeCard'
+import BadgeIcon from 'icons/BadgeIcon'
+import BadgeWrapper from 'components/badges/BadgeWrapper'
+import Button from 'components/Button'
+import ContractName from 'components/ContractName'
+import ExternalLink from 'components/ExternalLink'
+import getEtherscanAddressUrl from 'helpers/getEtherscanAddressUrl'
+
+export default function ({ contractAddress }: { contractAddress: string }) {
+  return (
+    <BadgeWrapper minted={false}>
+      <BadgeCard
+        top={<BadgeIcon />}
+        leanLeft
+        text={
+          <ExternalLink url={getEtherscanAddressUrl(contractAddress)}>
+            <ContractName address={contractAddress} />
+          </ExternalLink>
+        }
+        bottom={
+          <Button small primary loading>
+            Fetching...
+          </Button>
+        }
+      />
+    </BadgeWrapper>
+  )
+}
