--- src/models/WalletsToNotifiedOfBeingDoxxed.ts
@@ -1,3 +1,0 @@
-export default interface WalletsToNotifiedOfBeingDoxxed {
-  [address: string]: boolean
-}
