+++ src/stores/EnsStore.ts
@@ -9,6 +9,9 @@ interface EnsStoreInterface {
 const EnsStore = proxy<EnsStoreInterface>({
   ensNames: {},
   fetchEnsName(address: string) {
+    if (EnsStore.ensNames[address]) {
+      return
+    }
     EnsStore.ensNames[address] = defaultProvider.lookupAddress(address)
   },
 })
